refactor: use common components
