+++ apps/dashboard/app/(app)/audit/components/table/log-details/components/log-footer.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { RequestResponseDetails } from "@/app/(app)/logs/components/table/log-details/components/request-response-details";
+import { RequestResponseDetails } from "@/components/logs/details/request-response-details";
 import { TimestampInfo } from "@/components/timestamp-info";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import type { AuditLog } from "@/lib/trpc/routers/audit/schema";
