+++ apps/dashboard/app/(app)/audit/components/table/log-details/components/log-header.tsx
@@ -12,7 +12,7 @@ export const LogHeader = ({ onClose, log }: Props) => {
   return (
     <div className="border-b-[1px] flex justify-between items-center border-gray-4 pb-3 w-full">
       <div className="flex gap-2 items-center flex-1 min-w-0">
-        <Badge className="uppercase px-[6px] rounded-md font-mono bg-accent-3 text-accent-11 hover:bg-accent-4">
+        <Badge className="uppercase px-[6px] rounded-md font-mono bg-accent-4 text-accent-11 hover:bg-accent-5">
           {log.auditLog.event}
         </Badge>
       </div>
